+++ resources/[dg]/dg-misc/typescript/src/server/services/parachute.ts
@@ -1,0 +1,5 @@
+import { Events, Inventory } from '@dgx/server';
+
+Inventory.registerUseable('parachute', plyId => {
+  Events.emitNet('misc:parachute:toggle', plyId);
+});